Changes ImageConfiguration to be nested.
